Extract chatapp runtime event sink
